pyproject.toml
[project]
name = "financial-knowledge-graphs"
version = "0.1.0"
description = "Financial Knowledge Graphs"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "boto3==1.37.8",
    "botocore==1.37.8",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "distro==1.9.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.8.2",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core==0.3.41",
    "langchain-openai>=0.3.7",
    "langchain-text-splitters==0.3.6",
    "langsmith==0.3.12",
    "neo4j>=5.28.1",
    "openai==1.65.4",
    "orjson==3.10.15",
    "owlready2>=0.47",
    "packaging==24.2",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "s3transfer==0.11.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "stanford-openie>=1.3.2",
    "tenacity==9.0.0",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "zstandard==0.23.0",
]
